refactor: MAX_DEPTH is now u8
Since it's a constant, 100, why do I need the whole 64-bit (or 32-bit, or...)
number space? I don't.

+++ src/cli.rs
@@ -19,7 +19,7 @@ use serde::Deserialize;
 
 /// The maximum number of directories `agenix` is allowed to ascend in search of
 /// the `.agenix.toml` configuration.
-const MAX_DEPTH: usize = 100;
+const MAX_DEPTH: u8 = 100;
 
 #[doc(hidden)]
 const LF: [u8; 1] = [0x0a];
